termview/keymap: get the vim key name instead of a formatted one
The keymap was returning pre-formatted stringshare (e.g. F1 -> <F1>).
This was useful at some point when nvim_input() received stringshares,
but now it is much more flexible, so the keymap will just return the
official vim key name (plus its size), and formatting will be up to the
caller. This allows to handle composition better as well!