getopts: add support for -o, -O and -p
This is a bit tricky, as vim uses non-standard stuff to handle these.
Giving `-p a b`, `-p` would always consider that `a` is its argument. To
overcome this difference, -o, -O and -p take NO argument, are
equivalent to the vim options. If an argument is to be provided to
these, long options should be used instead: --hsplit, --vsplit and
--tabs.

Fixes #1