gui,theme: don't make the theme blink on its own
The guicursor option can configure how the cursor blinks. It is
therefore not up to the theme to decide of the blinking. The cursor's
theme has been rewritten to receive parameters from the gui, which are
now handled via Embryo scripts.

Partially addresses #4, as color is still not taken into account.