helper: use comand_output rather than eval to determine bg
Neovim 0.2.0 seems to have trouble with the eval RPC. This patch uses
the command_output RPC instead, hoping that it will work better. We now
only take care of the cursor background, as we do not implement the
foreground (modification of the text behind the cursor).

Hopefully, this should fix #8.