gui: fix resizing issues
This was a bit tough! First, the textgrid was resized manually. It
should not: the layout it supposed to do that automatically. This was
causing the mouse not to work fine.
Then, resizing was absolutely terrible. neovim was driving the UI
dimensions, but it was the UI that drove neovim to resize itself!

Now:
- when the UI is resized, we recalculate the cols/rows count, and ask
neovim to resize itself.
- meanwhile, we do not want to process obsolete data that neovim sent
before the resizing request. As such we now have a contract with neovim:
we use the signal send when it has been resized, so we now how to filter
old events from the relevant ones.

Fixes #6