nvim: place the commands under their method namespace
Neovim events are methods and commands. We previously acknowledged the
name of the method ("redraw") but didn't do anything about it (because
there is no need to). But this is not correct, as there can be several
methods with the same command names. As such, we must first match the
method name and then fin dout which function call for each command.

This opens the door for the Eovim API.