termview: fix size jitter on first sizing
At startup, the EFL window is resized several times. There is first a
garbage resizing with a height of 0, then the actual resizing with the
expected dimensions. But the termview was not resized with a strict
size, only with some size at some point, that could change upon
recalculating the window's constraints. Hence, the termview would shift
its size down, because the constraint of the window's size has a higher
priority than its components.

We now enforce a strict minimum size upon sizing, so the termview cannot
be re-evaluated with a dimension less than the one expected. This
correction of the sizing makes safety checks not to be triggered, since
nothing wrong should happen anymore at this stage.

Fixes #16.