termview: fix mouse dragging
Dragging with the mouse would select words instead of characters. This
was because Eovim sent too many signals to nvim. Actually, if two
MouseDrag signals are sent from the same cell, this is considered as a
"double click", and selects the whole word. Hence, the erratic behavior
of the mouse selection until now.