cmake: add the short git hash to eovim's version
In release (stable), version is pure semver (x.y.z). In dev mode
(nightly), a tweak version .99 is added. When the sources come from a
git repository, the git short hash it appended to this version.

This way, we are sure to always be able to make the distinction between
stable and dev, even though when eovim is compiled from outside of the
main git tree.

This also makes issue tracking much easier.