gui: fix invalid handling of group parts in edje
For some reason, there was a behavior change between efl 1.20.3 that
makes the edje theme behave as I was expected in latest EFL, but was
completely broken in efl <= 1.20.3 (maybe 1.20.4 also? untested).

We now stop relying on an undefined behavior and actually do handle the
swallowed text parts as they should.

Fixes #20