eovim: add caps lock auto-detection
Eovim now detects when the Caps Lock key is set/unset. The theme can
alert the user when Caps Lock is on, as this is mostly an annoyance. If
this feature is not appreciated, it can be disabled via the settings
panel.

More interesting, Eovim triggers a custom autocmd in Neovim when the
Caps Lock is on/off. See the README.md for details. You should setup
something like:

   :augroup EovimCapsLock
   :   autocmd!
   :   autocmd User EovimCapsLockOn :echo "hi caps lock"
   :   autocmd User EovimCapsLockOff :echo "bye caps lock"
   :augroup END

Fixes #27