nvim: source a dedicated vim config at startup
When eovim starts, it now sources an additional init script. There is no
guarantee WHEN this file will be loaded, but it is likely to be sourced
AFTER the initial nvim.init.

Use this file to include eovim-specific configuration code. It can be
either specified on the command-line or via the usual GUI config. It
defaults to ~/.config/nvim/eovimrc.vim.

Fixes #24.