eovim: refactorize events
nvim_event.c was too big of a file. To prepare the 'options_set' event
of the 'redraw' method, large events will be stored in their dedicated
file in the src/event/ directory.

This also marks the end of mode.c and mode.h, which are now contained
within the event that sets them.