event: don't spam the stderr with unimplemented details
Neovim sends to the GUI clients methods. Methods contain commands. The
GUI client may not be happy if:

- the data coming from neovim is ill-formed; or
- eovim did not keep up with neovim's features (or is not in sync).

In both cases, we printed an error, which by default would be displayed
in stderr. It has been lowered to a simple warning (not shown by
default). A normal user should not care. Increasing the log level will
show their the message.