cmdline: display user-defined prompts
Some plugins (e.g. vim-plug) display user-defined prompts. Eovim did not
display them, making it difficult for the user to understand what a vim
plugin was prompting.

In case of a user-defined prompt, the prompt will be displayed on TOP
of the text entry: it allows more room for the text to be displayed.