cmake: fix EFL detection
If seems that with more recent EFL, edje is not properly detected. Since
in the pkf-config file `-lm` comes before `-ledje`, the default cmake
files provided by EFL seem to detect edje as libm... which is obviously
not what is expected.

Instead of fighting with the EFL-distributed files, I prefered to make
my own, that are immune to this problem.