overhaul text rendering
Until now, Eovim was using the Evas_Textgrid to display text. It
consists in a fixed-size grid of codepoints associated with unique
glyphs and render attributes, such as the color and basic shape.

This has several problems, mainly:

- the textgrid has limited attributes support:
  - the color of underlined text is the same as the foreground
  - there is no "undercurl" equivalent
- you cannot specify a 'linespace' (#34)
- support for multiple width character is partial
- you cannot use the opentype capabilities such as ligatures (#44)

It was time for a change to support the above. Eovim now uses
Evas_Textblock, which is low-level rich text widget provided by EFL.
However, it is tricky to use the same way than the Textgrid, but a
proper separation between model/view/controller (wow... so innovative)
actually makes surprisingly it trivial to handle!

Fix #34
Fix #44