gui: overhaul cmdline and cursor
Cursor is now an autonomous entity, distinct from the termview. It also
replaces the ad-hoc edje cursor in the cmdline.

Command-line UI can not be customized through highlight groups, that are
properly detected at run-time.

Since this commit, we rely on user styles instead of text classes and
color classes for modifying the text attributes of the  cmdline.  This
also fixes a segfault caused by a bug in EFL 1.23.

Fix #52